Implement event editing and deletion features with Redux integration

+++ app/api/schedule/route.ts
@@ -27,7 +27,21 @@ export const GET = async (req: Request) => {
 
 export const POST = async (req: Request) => {
   const body = await req.json();
-  const { data, error } = await supabase.from("Schedule").insert([body]);
+  const { user_id, schedule_data, id } = body;
+  if (!user_id) {
+    return NextResponse.json(
+      { error: "Missing user_id in request body" },
+      { status: 400 }
+    );
+  }
+
+  const { data, error } = await supabase.from("Schedule").insert([
+    {
+      user_id: user_id,
+      schedule_data: schedule_data,
+      id: id,
+    },
+  ]);
 
   if (error) {
     console.error(error.message);
@@ -59,16 +73,41 @@ export const PATCH = async (req: Request) => {
 };
 
 export const DELETE = async (req: Request) => {
-  const { id } = await req.json();
-  const { data, error } = await supabase
-    .from("Schedule")
-    .delete()
-    .match({ id });
+  const { searchParams } = new URL(req.url);
+  const id = searchParams.get("id");
+  const deleteType = searchParams.get("deleteType");
+  if (!id || !deleteType) {
+    return NextResponse.json(
+      { error: "Missing id or deleteType parameter" },
+      { status: 400 }
+    );
+  }
+  if (deleteType === "single") {
+    const { data, error } = await supabase
+      .from("Schedule")
+      .delete()
+      .match({ id: id });
 
-  if (error) {
-    console.error(error.message);
-    return NextResponse.json(error.message);
-  } else {
-    return NextResponse.json(data);
+    if (error) {
+      console.error(error.message);
+      return NextResponse.json(error.message);
+    } else {
+      return NextResponse.json(data);
+    }
+  } else if (deleteType === "group") {
+    // delete whole group of recurring schedules
+    // column schedule_data is a json column
+    // one of the keys in the json is recurringGroupId
+    const { data, error } = await supabase
+      .from("Schedule")
+      .delete()
+      .match({ "schedule_data->>recurringGroupId": id });
+
+    if (error) {
+      console.error(error.message);
+      return NextResponse.json(error.message);
+    } else {
+      return NextResponse.json(data);
+    }
   }
 };
